Add fix for NaN rating showing for 0 rating gifts and optimized code structure
